Cleanup code in searchPage
