+++ frontend_nlp_react/src/taskpane/components/SearchPageComponents/SearchPage.jsx
@@ -3,7 +3,7 @@ import React, { useState } from "react";
 import InstructionPage from "./InstructionPage";
 import ResultsPage from "./ResultsPage";
 import FiltersPage from "./FiltersPage";
-import {useStyles, tabs, instructionPages, usStates, legislativeDocumentTypes} from "./SearchPageConsts"
+import {useStyles, tabs, instructionPages, usStates, legislativeDocumentTypes, MultiselectWithTags} from "./SearchPageConsts"
 
 const SearchPage = () => {
   // Page styling:
@@ -63,217 +63,6 @@ const SearchPage = () => {
   // Debugging
   const debug = () => {
     return `Output: ${selectedFileTypes}`;
-  };
-
-  // Complete Remove and Fix
-  const MultiselectWithTags = (props) => {
-    // generate ids for handling labelling
-    const comboId = useId("combo-multi");
-    const selectedListId = `${comboId}-selection`;
-
-    // refs for managing focus when removing tags
-    const selectedListRef = useRef(null);
-    const comboboxInputRef = useRef(null);
-
-    const options = usStates.map((state) => state.text);
-    const styles = useStyles();
-
-    // Handle selectedOptions both when an option is selected or deselected in the Combobox,
-    // and when an option is removed by clicking on a tag
-    const [selectedOptions, setSelectedOptions] = useState([]);
-
-    const onSelect = (event, data) => {
-      setSelectedOptions(data.selectedOptions);
-    };
-
-    const onTagClick = (option, index) => {
-      // remove selected option
-      setSelectedOptions(selectedOptions.filter((o) => o !== option));
-
-      // focus previous or next option, defaulting to focusing back to the combo input
-      const indexToFocus = index === 0 ? 1 : index - 1;
-      const optionToFocus = selectedListRef.current?.querySelector(`#${comboId}-remove-${indexToFocus}`);
-      if (optionToFocus) {
-        optionToFocus.focus();
-      } else {
-        comboboxInputRef.current?.focus();
-      }
-    };
-
-    const labelledBy = selectedOptions.length > 0 ? `${comboId} ${selectedListId}` : comboId;
-
-    return (
-      <div className={styles.root}>
-        {selectedOptions.length ? (
-          <ul id={selectedListId} className={styles.tagsList} ref={selectedListRef}>
-            {/* The "Remove" span is used for naming the buttons without affecting the Combobox name */}
-            <span id={`${comboId}-remove`} hidden>
-              Remove
-            </span>
-            {selectedOptions.map((option, i) => (
-              <li key={option}>
-                <Button
-                  size="small"
-                  shape="circular"
-                  appearance="primary"
-                  icon={<Dismiss12Regular />}
-                  iconPosition="after"
-                  onClick={() => onTagClick(option, i)}
-                  id={`${comboId}-remove-${i}`}
-                  aria-labelledby={`${comboId}-remove ${comboId}-remove-${i}`}
-                >
-                  {option}
-                </Button>
-              </li>
-            ))}
-          </ul>
-        ) : null}
-        <Combobox
-          aria-labelledby={labelledBy}
-          multiselect={true}
-          placeholder="Select States"
-          selectedOptions={selectedOptions}
-          onOptionSelect={onSelect}
-          ref={comboboxInputRef}
-          {...props}
-        >
-          {options.map((option) => (
-            <option key={option}>{option}</option>
-          ))}
-        </Combobox>
-      </div>
-    );
-  };
-
-
-  // Filter Tab Functions to handle interaction with filters.
-  // Handles applying file type
-  const handleFileTypeChange = (fileType) => {
-    setSelectedFileTypes((prevFileTypes) => {
-      if (prevFileTypes.includes(fileType)) {
-        return prevFileTypes.filter((type) => type !== fileType);
-      } else {
-        return [...prevFileTypes, fileType];
-      }
-    });
-  };
-
-  // Handles apply state
-  const handleStateChange = (event, option) => {
-    setSelectedState(option.text);
-  };
-
-  const onCheckboxChange = (ev, isChecked) => {
-    const fileType = ev.target.id;
-    handleFileTypeChange(fileType);
-  };
-
-  const onDropdownChange = (event, option, index) => {
-    handleStateChange(event, option);
-  };
-
-  const onDateSelect = (date) => {
-    setSelectedDate(date);
-  };
-
-  const handleRemoveFilter = (filterType) => {
-    switch (filterType) {
-      case "date":
-        setSelectedDate(null);
-        break;
-      case "fileType":
-        // setSelectedFileTypes([]);
-        break;
-      case "state":
-        setSelectedState(null);
-        break;
-      default:
-        break;
-    }
-  };
-
-  const renderChips = () => {
-    const [hoverStates, setHoverStates] = React.useState({
-      hover1: false,
-      hover2: false,
-      hover3: false,
-    });
-
-    const handleHover = (buttonIndex, isHovering) => {
-      setHoverStates((prevStates) => ({
-        ...prevStates,
-        [`hover${buttonIndex}`]: isHovering,
-      }));
-    };
-
-    const chipStyle = {
-      width: "auto",
-    };
-
-    const chips = [];
-
-    if (selectedDate) {
-      chips.push(
-        <ToggleButton
-          key="date"
-          size="small"
-          style={chipStyle}
-          onMouseOver={() => handleHover(1, true)}
-          onMouseOut={() => handleHover(1, false)}
-          {...(hoverStates.hover1 && { icon: <Dismiss24Regular /> })}
-          iconPosition="after"
-          shape="circular"
-          onClick={() => handleRemoveFilter("date")}
-        >
-          {selectedDate.toDateString()}
-        </ToggleButton>
-      );
-    }
-
-    if (selectedFileTypes.length > 0) {
-      selectedFileTypes.forEach((fileType, index) => {
-        chips.push(
-          <ToggleButton
-            key="date"
-            size="small"
-            style={chipStyle}
-            onMouseOver={() => handleHover(1, true)}
-            onMouseOut={() => handleHover(1, false)}
-            {...(hoverStates.hover1 && { icon: <Dismiss24Regular /> })}
-            iconPosition="after"
-            shape="circular"
-            onClick={() => handleRemoveFilter("fileType")}
-          >
-            {fileType}
-          </ToggleButton>
-        );
-      });
-    }
-
-    if (selectedState) {
-      chips.push(
-        <ToggleButton
-          key="date"
-          size="small"
-          style={chipStyle}
-          onMouseOver={() => handleHover(1, true)}
-          onMouseOut={() => handleHover(1, false)}
-          {...(hoverStates.hover1 && { icon: <Dismiss24Regular /> })}
-          iconPosition="after"
-          shape="circular"
-          onClick={() => handleRemoveFilter("state")}
-        >
-          {selectedState}
-        </ToggleButton>
-      );
-    }
-
-    return chips;
-  };
-
-  // Function to update chips
-  const updateChips = (newChips) => {
-    setChips(newChips);
   };
   
   return (
